chore(sourcepp): merge all math code into one header
