+++ include/sourcepp/crypto/Adler32.h
@@ -3,7 +3,7 @@
 #include <cstddef>
 #include <vector>
 
-#include <sourcepp/math/Integer.h>
+#include <sourcepp/Math.h>
 
 namespace sourcepp::crypto {
