+++ include/sourcepp/crypto/MD5.h
@@ -4,7 +4,7 @@
 #include <cstddef>
 #include <vector>
 
-#include <sourcepp/math/Integer.h>
+#include <sourcepp/Math.h>
 
 namespace sourcepp::crypto {
