+++ include/toolpp/CmdSeq.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include <sourcepp/math/Integer.h>
+#include <sourcepp/Math.h>
 
 namespace toolpp {
