+++ include/vcryptpp/VFONT.h
@@ -5,7 +5,7 @@
 #include <string_view>
 #include <vector>
 
-#include <sourcepp/math/Integer.h>
+#include <sourcepp/Math.h>
 
 namespace vcryptpp::VFONT {
