+++ include/vpkpp/Entry.h
@@ -5,7 +5,7 @@
 #include <variant>
 #include <vector>
 
-#include <sourcepp/math/Integer.h>
+#include <sourcepp/Math.h>
 
 namespace vpkpp {
