+++ src/sourcepp/_sourcepp.cmake
@@ -4,15 +4,11 @@ list(APPEND ${PROJECT_NAME}_HEADERS
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/crypto/MD5.h"
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/crypto/RSA.h"
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/crypto/String.h"
-        "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/math/Angles.h"
-        "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/math/Float.h"
-        "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/math/Integer.h"
-        "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/math/Matrix.h"
-        "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/math/Vector.h"
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/parser/Binary.h"
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/parser/Text.h"
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/FS.h"
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/Macros.h"
+        "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/Math.h"
         "${CMAKE_CURRENT_SOURCE_DIR}/include/sourcepp/String.h")
 
 add_library(${PROJECT_NAME} STATIC
